partway through refactoring again
- refactored and rearranged a little

# problem of bad terrain gen on game start persists
