--- necosystem/Assets/scripts/endlessTerrain.cs
@@ -1,291 +1,0 @@
-using JetBrains.Annotations;
-using System.Collections;
-using System.Collections.Generic;
-using Unity.VisualScripting;
-using UnityEngine;
-
-public class endlessTerrain : MonoBehaviour
-{
-    const float viewerMoveThresholdForChunkUpdate = 25f;
-    const float sqrViewerMoveThresholdForChunkUpdate = viewerMoveThresholdForChunkUpdate * viewerMoveThresholdForChunkUpdate;
-    const float colliderGenerationDistanceThreshold = 5;
-
-    public int colliderLODIndex;
-    public LODInfo[] detailLevels;
-    public static float maxViewDst;
-
-    public Transform viewer;
-    public Material mapMaterial;
-
-    public static Vector2 viewerPosition;
-    Vector2 viewerPositionOld;
-    static mapGenerator mapGenerator;
-    float meshWorldSize;
-    int chunksVisibleInViewDst;
-
-    Dictionary<Vector2, TerrainChunk> terrainChunkDictionary = new Dictionary<Vector2, TerrainChunk>();
-    static List<TerrainChunk> visibleTerrainChunks = new List<TerrainChunk>();
-
-    void Start()
-    {
-        mapGenerator = FindObjectOfType<mapGenerator>();
-
-        maxViewDst = detailLevels[detailLevels.Length - 1].visibleDstThreshold;
-        meshWorldSize = mapGenerator.meshSettings.meshWorldSize - 1;
-        chunksVisibleInViewDst = Mathf.RoundToInt(maxViewDst / meshWorldSize); //calculates visible chunks
-
-        UpdateVisibleChunks();
-    }
-
-    private void Update()
-    {
-        viewerPosition = new Vector2(viewer.position.x, viewer.position.z);
-
-        if (viewerPosition != viewerPositionOld)
-        {
-            foreach (TerrainChunk chunk in visibleTerrainChunks)
-            {
-                chunk.UpdateCollisionMesh();
-            }
-        }
-
-        if ((viewerPositionOld - viewerPosition).sqrMagnitude > sqrViewerMoveThresholdForChunkUpdate) //updates chunks if viewer has moved
-        {
-            viewerPositionOld = viewerPosition;
-            UpdateVisibleChunks();
-        }
-
-    }
-
-    void UpdateVisibleChunks()
-    {
-        HashSet<Vector2> alreadyUpdatedChunkCoords = new HashSet<Vector2>();
-        for (int i = visibleTerrainChunks.Count - 1; i >= 0; i--)
-        {
-            alreadyUpdatedChunkCoords.Add(visibleTerrainChunks[i].coord);
-            visibleTerrainChunks[i].updateTerrainChunk(); //clear up chunks that should no longer be visible
-        }
-
-        int currentChunkCoordX = Mathf.RoundToInt(viewerPosition.x / meshWorldSize);
-        int currentChunkCoordY = Mathf.RoundToInt(viewerPosition.y / meshWorldSize);
-
-        for (int yOffset = -chunksVisibleInViewDst; yOffset <= chunksVisibleInViewDst; yOffset++)
-            for (int xOffset = -chunksVisibleInViewDst; xOffset <= chunksVisibleInViewDst; xOffset++)
-            { 
-                Vector2 viewedChunkCoord = new Vector2(currentChunkCoordX + xOffset, currentChunkCoordY + yOffset);
-
-                if (!alreadyUpdatedChunkCoords.Contains(viewedChunkCoord))
-                {
-                    if (terrainChunkDictionary.ContainsKey(viewedChunkCoord))
-                    {
-                        terrainChunkDictionary[viewedChunkCoord].updateTerrainChunk();
-
-                    }
-                    else
-                    {
-                        terrainChunkDictionary.Add(viewedChunkCoord, new TerrainChunk(viewedChunkCoord, meshWorldSize, detailLevels, colliderLODIndex, transform, mapMaterial)); //add the loaded chunk into the dictionary
-                    }
-                }
-            }
-    }
-
-    public class TerrainChunk
-    {
-        public Vector2 coord;
-
-        GameObject meshObject;
-        Vector2 sampleCentre;
-        Bounds bounds;
-
-        MeshRenderer meshRenderer;
-        MeshFilter meshFilter;
-        MeshCollider meshCollider;
-
-        LODInfo[] detailLevels;
-        LODMesh[] lodMeshes;
-        int colliderLODIndex;
-
-        HeightMap mapData;
-        bool mapDataReceived;
-        int previousLODIndex = -1;
-        bool hasSetCollider;
-
-        public TerrainChunk(Vector2 coord, float meshWorldSize, LODInfo[] detailLevels, int colliderLODIndex, Transform parent, Material material)
-        {
-            this.coord = coord;
-            this.detailLevels = detailLevels;
-            this.colliderLODIndex = colliderLODIndex;
-
-            sampleCentre = coord * meshWorldSize / mapGenerator.meshSettings.meshScale;
-            Vector2 position = coord * meshWorldSize;
-            bounds = new Bounds(position, Vector2.one * meshWorldSize);
-
-            meshObject = new GameObject("Terrain Chunk");
-            meshRenderer = meshObject.AddComponent<MeshRenderer>();
-            meshFilter = meshObject.AddComponent<MeshFilter>();
-            meshCollider = meshObject.AddComponent<MeshCollider> ();
-            meshRenderer.material = material;
-
-            meshObject.transform.position = new Vector3(position.x,0,position.y);
-            meshObject.transform.parent = parent;
-
-            SetVisible(false); //defaults chunks to not visible
-
-            lodMeshes = new LODMesh[detailLevels.Length];
-            for (int i = 0; i < detailLevels.Length; i++)
-            {
-                lodMeshes[i] = new LODMesh(detailLevels[i].lod);
-                lodMeshes[i].updateCallback += updateTerrainChunk;
-                if (i == colliderLODIndex)
-                {
-                    lodMeshes[i].updateCallback += UpdateCollisionMesh;
-                }
-            }
-
-            mapGenerator.RequestHeightMap(sampleCentre, OnMapDataReceived);
-        }
-
-        void OnMapDataReceived(HeightMap mapData)
-        {
-            this.mapData = mapData;
-            mapDataReceived = true;
-
-            updateTerrainChunk();
-        }
-
-        public void updateTerrainChunk() //sets the LOD value of a chunk depending on distance
-        {
-            if (mapDataReceived)
-            {
-                float viewerDstFromNearestEdge = Mathf.Sqrt(bounds.SqrDistance(viewerPosition)); //calculates the distance between viewer and chunk
-
-                bool wasVisible = isVisible();
-                bool visible = viewerDstFromNearestEdge <= maxViewDst; //determines if its visible based on distance
-
-                if (visible)
-                {
-                    int lodIndex = 0;
-
-                    for (int i = 0; i < detailLevels.Length - 1; i++)
-                    {
-                        if (viewerDstFromNearestEdge > detailLevels[i].visibleDstThreshold)
-                        {
-                            lodIndex = i + 1;
-                        }
-                        else
-                        {
-                            break;
-                        }
-                    }
-
-                    if (lodIndex != previousLODIndex)
-                    {
-                        LODMesh lodMesh = lodMeshes[lodIndex];
-                        if (lodMesh.hasMesh)
-                        {
-                            previousLODIndex = lodIndex;
-                            meshFilter.mesh = lodMesh.mesh;
-                        }
-                        else if (!lodMesh.hasRequestedMesh)
-                        {
-                            lodMesh.RequestMesh(mapData);
-                        }
-                    }
-
-                    
-                }
-
-                if (wasVisible != visible)
-                {
-                    if (visible)
-                    {
-                        visibleTerrainChunks.Add(this); //adds the currently visible chunk to the list
-                    }
-                    else
-                    {
-                        visibleTerrainChunks.Remove(this);
-                    }
-                    SetVisible(visible);
-                }
-                
-            }
-        }
-
-        public void UpdateCollisionMesh()
-        {
-            if (!hasSetCollider)
-            {
-                float sqrDstFromViewerToEdge = bounds.SqrDistance(viewerPosition);
-
-                if (sqrDstFromViewerToEdge < detailLevels[colliderLODIndex].sqrVisibleDstThreshold)
-                {
-                    if (!lodMeshes[colliderLODIndex].hasRequestedMesh)
-                    {
-                        lodMeshes[colliderLODIndex].RequestMesh(mapData);
-                        hasSetCollider = true;
-                    }
-                }
-
-                if (sqrDstFromViewerToEdge < colliderGenerationDistanceThreshold * colliderGenerationDistanceThreshold)
-                {
-                    if (lodMeshes[colliderLODIndex].hasMesh)
-                        meshCollider.sharedMesh = lodMeshes[colliderLODIndex].mesh;
-                }
-            }
-        }
-
-        public void SetVisible(bool visible)
-        {
-            meshObject.SetActive(visible);
-        }
-
-        public bool isVisible()
-        {
-            return meshObject.activeSelf;
-        }
-    }
-
-    class LODMesh
-    {
-        public Mesh mesh;
-        public bool hasRequestedMesh;
-        public bool hasMesh;
-        int lod;
-        public event System.Action updateCallback;
-
-        public LODMesh(int lod)
-        {
-            this.lod = lod;
-        }
-
-        void OnMeshDataReceived(MeshData meshData)
-        {
-            mesh = meshData.CreateMesh();
-            hasMesh = true;
-
-            updateCallback();
-        }
-
-        public void RequestMesh(HeightMap mapData)
-        {
-            hasRequestedMesh = true;
-            mapGenerator.RequestMeshData(mapData, lod, OnMeshDataReceived);
-        }
-    }
-
-    [System.Serializable]
-    public struct LODInfo
-    {
-        [Range(0,MeshSettings.numSupportedLODs-1)]
-        public int lod;
-        public float visibleDstThreshold;
-
-        public float sqrVisibleDstThreshold
-        {
-            get
-            {
-                return visibleDstThreshold * visibleDstThreshold;
-            }
-        }
-    }
-}
