+++ necosystem/Assets/scripts/terrainChunk.cs
@@ -1,0 +1,207 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TerrainChunk
+{
+    const float colliderGenerationDistanceThreshold = 5;
+    public event System.Action<TerrainChunk, bool> onVisibilityChanged;
+    public Vector2 coord;
+
+    GameObject meshObject;
+    Vector2 sampleCentre;
+    Bounds bounds;
+
+    MeshRenderer meshRenderer;
+    MeshFilter meshFilter;
+    MeshCollider meshCollider;
+
+    LODInfo[] detailLevels;
+    LODMesh[] lodMeshes;
+    int colliderLODIndex;
+
+    HeightMap heightMap;
+    bool heightMapReceived;
+    int previousLODIndex = -1;
+    bool hasSetCollider;
+    float maxViewDst;
+
+    HeightMapSettings heightMapSettings;
+    MeshSettings meshSettings;
+    Transform viewer;
+
+    public TerrainChunk(Vector2 coord, HeightMapSettings heightMapSettings, MeshSettings meshSettings, LODInfo[] detailLevels, int colliderLODIndex, Transform parent, Transform viewer, Material material)
+    {
+        this.coord = coord;
+        this.detailLevels = detailLevels;
+        this.colliderLODIndex = colliderLODIndex;
+        this.heightMapSettings = heightMapSettings;
+        this.meshSettings = meshSettings;
+        this.viewer = viewer;
+
+        sampleCentre = coord * meshSettings.meshWorldSize / meshSettings.meshScale;
+        Vector2 position = coord * meshSettings.meshWorldSize;
+        bounds = new Bounds(position, Vector2.one * meshSettings.meshWorldSize);
+
+        meshObject = new GameObject("Terrain Chunk");
+        meshRenderer = meshObject.AddComponent<MeshRenderer>();
+        meshFilter = meshObject.AddComponent<MeshFilter>();
+        meshCollider = meshObject.AddComponent<MeshCollider>();
+        meshRenderer.material = material;
+
+        meshObject.transform.position = new Vector3(position.x, 0, position.y);
+        meshObject.transform.parent = parent;
+
+        SetVisible(false); //defaults chunks to not visible
+
+        lodMeshes = new LODMesh[detailLevels.Length];
+        for (int i = 0; i < detailLevels.Length; i++)
+        {
+            lodMeshes[i] = new LODMesh(detailLevels[i].lod);
+            lodMeshes[i].updateCallback += updateTerrainChunk;
+            if (i == colliderLODIndex)
+            {
+                lodMeshes[i].updateCallback += UpdateCollisionMesh;
+            }
+        }
+
+        maxViewDst = detailLevels[detailLevels.Length - 1].visibleDstThreshold;
+
+        
+    }
+
+    public void Load()
+    {
+        threadedDataRequester.RequestData(() => heightMapGenerator.generateHeightMap(meshSettings.numVertsPerLine, meshSettings.numVertsPerLine, heightMapSettings, sampleCentre), OnHeightMapReceived);
+    }
+
+    void OnHeightMapReceived(object heightMapObject)
+    {
+        this.heightMap = (HeightMap)heightMapObject;
+        heightMapReceived = true;
+
+        updateTerrainChunk();
+    }
+
+    Vector2 viewerPosition
+    {
+        get
+        {
+            return new Vector2(viewer.position.x, viewer.position.z);
+        }
+    }
+
+    public void updateTerrainChunk() //sets the LOD value of a chunk depending on distance
+    {
+        if (heightMapReceived)
+        {
+            float viewerDstFromNearestEdge = Mathf.Sqrt(bounds.SqrDistance(viewerPosition)); //calculates the distance between viewer and chunk
+
+            bool wasVisible = isVisible();
+            bool visible = viewerDstFromNearestEdge <= maxViewDst; //determines if its visible based on distance
+
+            if (visible)
+            {
+                int lodIndex = 0;
+
+                for (int i = 0; i < detailLevels.Length - 1; i++)
+                {
+                    if (viewerDstFromNearestEdge > detailLevels[i].visibleDstThreshold)
+                    {
+                        lodIndex = i + 1;
+                    }
+                    else
+                    {
+                        break;
+                    }
+                }
+
+                if (lodIndex != previousLODIndex)
+                {
+                    LODMesh lodMesh = lodMeshes[lodIndex];
+                    if (lodMesh.hasMesh)
+                    {
+                        previousLODIndex = lodIndex;
+                        meshFilter.mesh = lodMesh.mesh;
+                    }
+                    else if (!lodMesh.hasRequestedMesh)
+                    {
+                        lodMesh.RequestMesh(heightMap, meshSettings);
+                    }
+                }
+
+
+            }
+
+            if (wasVisible != visible)
+            {
+                SetVisible(visible);
+                if (onVisibilityChanged != null)
+                {
+                    onVisibilityChanged(this, visible);
+                }
+            }
+        }
+    }
+
+    public void UpdateCollisionMesh()
+    {
+        if (!hasSetCollider)
+        {
+            float sqrDstFromViewerToEdge = bounds.SqrDistance(viewerPosition);
+
+            if (sqrDstFromViewerToEdge < detailLevels[colliderLODIndex].sqrVisibleDstThreshold)
+            {
+                if (!lodMeshes[colliderLODIndex].hasRequestedMesh)
+                {
+                    lodMeshes[colliderLODIndex].RequestMesh(heightMap, meshSettings);
+                    hasSetCollider = true;
+                }
+            }
+
+            if (sqrDstFromViewerToEdge < colliderGenerationDistanceThreshold * colliderGenerationDistanceThreshold)
+            {
+                if (lodMeshes[colliderLODIndex].hasMesh)
+                    meshCollider.sharedMesh = lodMeshes[colliderLODIndex].mesh;
+            }
+        }
+    }
+
+    public void SetVisible(bool visible)
+    {
+        meshObject.SetActive(visible);
+    }
+
+    public bool isVisible()
+    {
+        return meshObject.activeSelf;
+    }
+}
+
+class LODMesh
+{
+    public Mesh mesh;
+    public bool hasRequestedMesh;
+    public bool hasMesh;
+    int lod;
+    public event System.Action updateCallback;
+
+    public LODMesh(int lod)
+    {
+        this.lod = lod;
+    }
+
+    void OnMeshDataReceived(object meshDataObject)
+    {
+        mesh = ((MeshData)meshDataObject).CreateMesh();
+        hasMesh = true;
+
+        updateCallback();
+    }
+
+    public void RequestMesh(HeightMap heightMap, MeshSettings meshSettings)
+    {
+        hasRequestedMesh = true;
+        threadedDataRequester.RequestData(() => meshGenerator.GenerateTerrainMesh(heightMap.values, meshSettings, lod), OnMeshDataReceived);
+    }
+}
