+++ necosystem/Assets/scripts/terrainChunk.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5fb2f20681fd6e5458f3e79e660c60a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
